refactor: 상수 값 안에서 초기화
