+++ src/main/java/racingcar/domain/Car/NumberGenerator.java
@@ -1,5 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.Car;
 
 public interface NumberGenerator {
+
     int generate();
 }
