+++ src/main/java/racingcar/domain/Car/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.Car;
 
 import org.springframework.stereotype.Component;
